refactor(utils): split internal and external tools
