+++ sdk/src/BookingAPI.test.ts
@@ -1,6 +1,6 @@
 import BookingAPI from './BookingAPI.mock';
 import { Booking, Resource, Schedule } from './types';
-import { openingHourGenerator } from './utils';
+import { openingHourGenerator } from './utils.internal';
 import {
   BadRequestError,
   ConflictingObjectExists,
